지표값을 double에서 long double로 확장

+++ Includes/Indicators/Open.hpp
@@ -13,5 +13,5 @@ class Open final : public Indicator {
   int symbol_idx_;
 
   void Initialize() override;
-  Numeric<double> Calculate() override;
+  Numeric<long double> Calculate() override;
 };
